requirements.txt
transformers==4.44.2
sentence-transformers==3.1.0
python-dotenv==1.0.1
chainlit==1.0.506
langsmith==0.1.121
langchain-huggingface==0.1.0
langchain-core==0.3.1
jsonpointer==3.0.0
huggingface-hub==0.25.0
torch==2.4.1
bitsandbytes==0.43.3
opentelemetry-api==1.27.0
numpy==1.26.4

-e .